+++ internal/constants/constants.go
@@ -3,8 +3,8 @@ package constants
 import "time"
 
 const (
-	Version              = "2.3.1"
-	VersionName          = "Ao"
+	Version              = "2.4.0"
+	VersionName          = "Aka"
 	GcTime               = time.Minute * 30
 	ConfigFileName       = "config.toml"
 	MalClientId          = "51cb4294feb400f3ddc66a30f9b9a00f"